refactor: [FE] 채팅 영역 렌더링 개선 React memo 적용 및 state 하위 컴포넌트로 이동 (#217)
